+++ apps/lite/ui/src/errors.ts
@@ -1,0 +1,10 @@
+export const errorMessageForToast = (error: unknown): string => {
+	if (error instanceof Error) return error.message;
+	if (typeof error === "string") return error;
+
+	try {
+		return JSON.stringify(error);
+	} catch {
+		return "Unknown error.";
+	}
+};
